+++ webapp/src/main/java/rocks/metaldetector/config/WebMvcConfig.java
@@ -8,9 +8,9 @@ import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import rocks.metaldetector.config.constants.Endpoints;
 import rocks.metaldetector.config.constants.ViewNames;
 import rocks.metaldetector.security.RedirectionHandlerInterceptor;
+import rocks.metaldetector.support.Endpoints;
 
 @Configuration
 public class WebMvcConfig implements WebMvcConfigurer {
