+++ webapp/src/main/java/rocks/metaldetector/security/handler/CustomAuthenticationSuccessHandler.java
@@ -4,7 +4,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
 import org.springframework.security.web.authentication.SavedRequestAwareAuthenticationSuccessHandler;
 import org.springframework.security.web.savedrequest.DefaultSavedRequest;
-import rocks.metaldetector.config.constants.Endpoints;
+import rocks.metaldetector.support.Endpoints;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
