+++ webapp/src/main/java/rocks/metaldetector/web/controller/mvc/CustomErrorController.java
@@ -6,8 +6,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
-import rocks.metaldetector.config.constants.Endpoints;
 import rocks.metaldetector.config.constants.ViewNames;
+import rocks.metaldetector.support.Endpoints;
 
 import javax.servlet.RequestDispatcher;
 import javax.servlet.http.HttpServletRequest;
