+++ webapp/src/main/java/rocks/metaldetector/web/controller/mvc/MyArtistsController.java
@@ -4,8 +4,8 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
-import rocks.metaldetector.config.constants.Endpoints;
 import rocks.metaldetector.config.constants.ViewNames;
+import rocks.metaldetector.support.Endpoints;
 
 @Controller
 @RequestMapping(Endpoints.Frontend.MY_ARTISTS)
