+++ webapp/src/main/java/rocks/metaldetector/web/controller/mvc/admin/AdminIndexController.java
@@ -3,8 +3,8 @@ package rocks.metaldetector.web.controller.mvc.admin;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.servlet.ModelAndView;
-import rocks.metaldetector.config.constants.Endpoints;
 import rocks.metaldetector.config.constants.ViewNames;
+import rocks.metaldetector.support.Endpoints;
 
 @Controller
 public class AdminIndexController {
