+++ webapp/src/test/java/rocks/metaldetector/security/handler/CustomAuthenticationFailureHandlerTest.java
@@ -12,7 +12,7 @@ import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.DisabledException;
 import org.springframework.security.authentication.InternalAuthenticationServiceException;
 import org.springframework.security.web.authentication.AuthenticationFailureHandler;
-import rocks.metaldetector.config.constants.Endpoints;
+import rocks.metaldetector.support.Endpoints;
 
 class CustomAuthenticationFailureHandlerTest implements WithAssertions {
