+++ webapp/src/test/java/rocks/metaldetector/web/controller/rest/MyArtistsRestControllerTest.java
@@ -13,10 +13,10 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
-import rocks.metaldetector.config.constants.Endpoints;
 import rocks.metaldetector.service.SlicingService;
 import rocks.metaldetector.service.artist.ArtistDto;
 import rocks.metaldetector.service.artist.FollowArtistService;
+import rocks.metaldetector.support.Endpoints;
 import rocks.metaldetector.testutil.DtoFactory.ArtistDtoFactory;
 import rocks.metaldetector.web.RestAssuredMockMvcUtils;
 import rocks.metaldetector.web.api.response.MyArtistsResponse;
