+++ webapp/src/test/java/rocks/metaldetector/web/controller/rest/ReleasesRestControllerTest.java
@@ -19,9 +19,9 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import rocks.metaldetector.butler.facade.ReleaseService;
 import rocks.metaldetector.butler.facade.dto.ImportJobResultDto;
-import rocks.metaldetector.butler.facade.dto.ReleaseDto;
-import rocks.metaldetector.config.constants.Endpoints;
 import rocks.metaldetector.service.exceptions.RestExceptionsHandler;
+import rocks.metaldetector.butler.facade.dto.ReleaseDto;
+import rocks.metaldetector.support.Endpoints;
 import rocks.metaldetector.support.Page;
 import rocks.metaldetector.support.PageRequest;
 import rocks.metaldetector.support.Pagination;
