+++ webapp/src/test/java/rocks/metaldetector/web/controller/rest/SummaryRestControllerTest.java
@@ -10,8 +10,8 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import rocks.metaldetector.config.constants.Endpoints;
 import rocks.metaldetector.service.summary.SummaryService;
+import rocks.metaldetector.support.Endpoints;
 import rocks.metaldetector.testutil.DtoFactory.ReleaseDtoFactory;
 import rocks.metaldetector.web.RestAssuredMockMvcUtils;
 import rocks.metaldetector.web.api.response.SummaryResponse;
